// lib/model/pet.tsx
export interface Pet {
    petType: string;
    litter: string;
    litterDetails: string;
    sex: string;
    isFixed: string;
    isShowPet: string;
    furColor: string; 
    hypoAllergies: string;
    pricePoint: string;
    eyeColor: string;
    age: string;
    hairLength: string;
    comments: string;
}